Support base class and metaclass mode
